Added some error detection to daap reading/writing and fixed some incorrect codes. Also updated codes and added content-codes request

+++ trunk/DAAPLib/src/util/response/CtrlInt.java
@@ -71,7 +71,7 @@ public class CtrlInt extends Response {
 				}
 			});
 		}
-		writer.appendList(Constants.dmap_listing, 0, nodes);
+		writer.appendList(Constants.dmap_listing, (byte)0, nodes);
 	}
 	
 	public Iterable<Integer> ids() {
